feat: login from any page
