+++ src/routes/my-wishlist/+page.svelte
@@ -26,23 +26,6 @@
 	</div>
 </div>
 
-<!-- <div class="m-4 flex items-center justify-between"> -->
-<!-- <div class="m-4 flex justify-end">
-	<h1 class="text-xl font-bold w">My Wishlist</h1>
-	<button
-		class="btn btn-accent"
-		onclick={() => {
-			showWishModal = true;
-		}}>+ Add Wish</button
-	>
-
-	<div class="columns-sm space-y-5 p-5">
-		{#each data.wishes.map((wish, i) => ({ ...wish, shadowIndex: i % 4 })) as wish}
-			<WishCard {...wish} />
-		{/each}
-	</div>
-</div> -->
-
 {#if showWishModal}
 	<WishModal bind:openModal={showWishModal} />
 {/if}
